refactor(protocol-jackson): use protocol marker annotations
Instead of using the `protocol-util` module, use the protocol marker
annotations to correctly deserialize the protocol's abstract types into
concrete types. This will also allow deserializing the immutable classes
as well.
