+++ protocol-jackson/src/main/java/io/camunda/zeebe/protocol/jackson/RecordValueTypeIdResolver.java
@@ -9,8 +9,8 @@ package io.camunda.zeebe.protocol.jackson;
 
 import io.camunda.zeebe.protocol.record.RecordValue;
 import io.camunda.zeebe.protocol.record.ValueType;
-import io.camunda.zeebe.protocol.util.ValueTypeMapping;
-import io.camunda.zeebe.protocol.util.ValueTypeMapping.Mapping;
+import io.camunda.zeebe.protocol.record.ValueTypeMapping;
+import io.camunda.zeebe.protocol.record.ValueTypeMapping.Mapping;
 import java.util.Objects;
 
 /**
