+++ protocol-jackson/src/test/java/io/camunda/zeebe/protocol/jackson/IntentTypeIdResolverTest.java
@@ -14,8 +14,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.deser.BeanDeserializerFactory;
 import com.fasterxml.jackson.databind.deser.DefaultDeserializationContext;
 import io.camunda.zeebe.protocol.record.ValueType;
+import io.camunda.zeebe.protocol.record.ValueTypeMapping;
 import io.camunda.zeebe.protocol.record.intent.Intent;
-import io.camunda.zeebe.protocol.util.ValueTypeMapping;
 import java.io.IOException;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.parallel.Execution;
