+++ protocol-jackson/src/test/java/io/camunda/zeebe/protocol/jackson/ZeebeProtocolModuleTest.java
@@ -15,8 +15,8 @@ import io.camunda.zeebe.protocol.record.Record;
 import io.camunda.zeebe.protocol.record.RecordAssert;
 import io.camunda.zeebe.protocol.record.RecordValue;
 import io.camunda.zeebe.protocol.record.ValueType;
-import io.camunda.zeebe.protocol.util.ValueTypeMapping;
-import io.camunda.zeebe.protocol.util.ValueTypeMapping.Mapping;
+import io.camunda.zeebe.protocol.record.ValueTypeMapping;
+import io.camunda.zeebe.protocol.record.ValueTypeMapping.Mapping;
 import io.camunda.zeebe.test.broker.protocol.ProtocolFactory;
 import java.io.IOException;
 import java.util.stream.Stream;
